Add ServletContext to ContextImpl to improve 3rd-party integration
Some servlet-based integrations require access to the ServletContext. While
this could probably be faked with a custom wrapper around Ninja's Context
it is easier to just provide the actual context to ContextImpl for those integrations.

One example integration which would benefit from exposing the ServletContext
is the Thymeleaf template engine[1]. It features built-in servlet integration which
can not be used without a ServletContext.

- [1]: http://thymeleaf.org