Add new Ninja logo and improve typography & styling of documentation
I've created an SVG from the shape of the console ASCII logo and remixed it with a
Ninja star.  I also created a square logo for social media accounts, favicons, and
other purposes.

I believe the css styling tweaks, font changes, and introduction of FontAwesome
produce a better-formatted website.

Obviously all of this is subjective.