Revise the boolean setting controlling external configuration hot-reloading
The addition of the setting within the external application.conf file to control
enabling/disabling runtime hot-reloading for #235 was hastily done.

In retrospect, it would be better to control this behavior by use of a System
property which complements the system property already used to specify the
external configuration location.

    java -Dninja.external.reload=true -Dninja.external.configuration=app.conf