Re-throw template errors and buffer templates.
Added rethrowing of template errors in prod mode so the application can
return it's 500 code or otherwise handle the exception from the main
exception handling code. This requires full buffering of the template
output since we can't guarantee that we haven't flushed some part of the
response to the user. I think in most cases this will not matter, as the
performance tests I ran showed slightly higher GC pressure but a bit
more reqs/s with full buffering.