Fix bug in AssetsController in dev mode
Bug happened in dev mode on App Engine. Because
user.dir returns the war directory that does
not contain the src/main/assets folder it breaks
dev mode. This patch tests for the file and
falls back to loading from classpath if file
can not be loaded from file system.