Fix issue exposed when saving the cookie on every request.
Redirect routes were returning the same Result object every time and this
object could be modified by the session or by filters. The fix is to
make a copy of this Result on each request before it can be modified.

Since setting the cookie on every request is prone to race conditions,
dont do it in the default example and update the session docs to
explain the issue.