Fix issue exposed when saving the cookie on every request.
RouteBuilder.with(Result) style routes are now removed.  If you are
using any of these you will need to switch to a Controller route, like
so:

conf/Route.java:

  router.GET().route("/.*").with(ApplicationController.class, "redirector");

ApplicationController.java:

  public Result redirector() {
     return Results.redirect("/");
  }

Redirect routes were returning the same Result object every time and this
object could be modified by the session or by filters.

Also since setting the cookie on every request is prone to race conditions,
dont do it in the default example and update the session docs to explain
the issue.